fix: P3 audit findings — docs, observability, API design, robustness
31 fixes across 20 files from v0.12.12 audit P3 tier:

Documentation (9): README --expand flag, health/suggest/--include-types docs,
CONTRIBUTING file listings, CHANGELOG URLs, ROADMAP checkmarks, SECURITY
reranker model

Error handling (5): .context() on Store::open in drift, embed_query in batch,
debug→warn in staleness, EH-25/EH-29 documented/n/a

Observability (4): tracing spans for run_index_pipeline, apply_windowing,
GPU/CPU embed threads, AtomicU64 error counter in batch stats

API design (5): Debug+Clone on TypeGraph/ResolvedTarget, Serialize on Note,
re-export drift types, TypeEdgeKind enum replaces stringly-typed edge_kind

Robustness (5): onboard depth clamped to 10, search_by_name limit clamped
to 100, TypeGraph MAX_EDGES as usize, RB-25/RB-30 verified correct

Algorithm/platform/perf (5): onboard total_items includes key_types,
AC-27 numbering verified correct, PathBuf forward-slash serialization,
Cow<str> in strip_markdown_noise, PERF-26 verified correct

Test coverage (2): TypeEdgeKind round-trip test, staleness test assertion

8 findings were non-issues or already fixed in P2.

+++ src/store/types.rs
@@ -29,6 +29,7 @@ pub struct TypeEdgeStats {
 /// Built from a single scan of the `type_edges` table joined with `chunks`.
 /// Forward: chunk_name -> Vec<type_name>, Reverse: type_name -> Vec<chunk_name>.
 /// Used by Phase 4 BFS traversal over type edges.
+#[derive(Debug, Clone)]
 pub struct TypeGraph {
     /// Forward edges: chunk_name -> Vec<type_name>
     pub forward: HashMap<String, Vec<String>>,
@@ -414,18 +415,18 @@ impl Store {
         let _span = tracing::info_span!("get_type_graph").entered();
 
         self.rt.block_on(async {
-            const MAX_TYPE_GRAPH_EDGES: i64 = 500_000;
+            const MAX_TYPE_GRAPH_EDGES: usize = 500_000;
             let rows: Vec<(String, String)> = sqlx::query_as(
                 "SELECT c.name, te.target_type_name
                  FROM type_edges te
                  JOIN chunks c ON te.source_chunk_id = c.id
                  LIMIT ?1",
             )
-            .bind(MAX_TYPE_GRAPH_EDGES)
+            .bind(MAX_TYPE_GRAPH_EDGES as i64)
             .fetch_all(&self.pool)
             .await?;
 
-            if rows.len() as i64 >= MAX_TYPE_GRAPH_EDGES {
+            if rows.len() >= MAX_TYPE_GRAPH_EDGES {
                 tracing::warn!(
                     cap = MAX_TYPE_GRAPH_EDGES,
                     "Type graph hit edge cap, results may be incomplete"
